mitarbeiterController added
git-svn-id: file:///home/node/svn/project_se/goran/EPU_Office@134 bde5ac68-c2e4-480a-bd7c-0859b87ef666

+++ src/view/KontakteView.java
@@ -341,7 +341,9 @@ public final class KontakteView extends AbstractViewPanel {
     private void kontaktLoeschenActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_kontaktLoeschenActionPerformed
         Kontakt k = (Kontakt) this.kontaktListe.getSelectedValue();
         try {
+            if(!(kontaktListe.isSelectionEmpty())){
             db.deleteKontakt(k);
+            }
         } catch (DALException ex) {
             Logger.log(Level.SEVERE, KontakteView.class, ex);
         }
